+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 12/02/2021
+
+- Remove dependency on CT's `authenticated` functionality
+
 ## 19/11/2020
 
 - Fix issue where metadata status could not be updated.